Refactor: 컬렉션 디테일 페이지 컴포넌트 헤더 바디 병합
