+++ src/features/collection/ui/CollectionDetail.style.ts
