+++ src/pages/collection/CollectionDetailPage.tsx
@@ -1,12 +1,10 @@
-import { CollectionDetailBody } from 'features/collection/ui/CollectionDetailBody';
-import { CollectionHeader } from 'features/collection/ui/CollectionHeader';
+import { CollectionDetail } from 'features/collection/ui/CollectionDetail';
 import { Container } from 'shared/styles/Container';
 
 export default function CollectionDetailPage() {
   return (
     <Container>
-      <CollectionHeader />
-      <CollectionDetailBody />
+      <CollectionDetail />
     </Container>
   );
 }
